consolidated functions with board process6, merged collision files, named game idle file.
