+++ ld_script.txt
@@ -46,9 +46,8 @@ SECTIONS
         src/board_process5.o(.text);
         src/board_process7.o(.text);
         src/board_process6.o(.text);
-        src/rom_14740.o(.text);
-        src/rom_17634.o(.text);
-        src/rom_19028.o(.text);
+        src/collision_checks.o(.text);
+        src/game_idle.o(.text);
         src/board_process4.o(.text);
         src/board_process3_ruby.o(.text);
         src/rom_1A0F4.o(.text);
