+++ src/game_idle.c
